@@ -2,46 +2,39 @@
     "name": "laravel/laravel",
     "type": "project",
     "description": "The Laravel Framework.",
-    "keywords": [
-        "framework",
-        "laravel"
-    ],
+    "keywords": ["framework", "laravel"],
     "license": "MIT",
     "require": {
-        "php": "^7.3",
-        "fideloper/proxy": "^4.2",
+        "php": "7.4.*|^8.0",
+        "barryvdh/laravel-ide-helper": "^2.10",
+        "fideloper/proxy": "^4.4",
         "fruitcake/laravel-cors": "^2.0",
         "guzzlehttp/guzzle": "^7.0.1",
         "ixudra/curl": "^6.21",
-        "laravel/fortify": "^1.4",
-        "laravel/framework": "^8.0",
-        "laravel/tinker": "^2.0",
+        "knuckleswtf/scribe": "^2.6",
+        "laravel/framework": "^8.12",
+        "laravel/tinker": "^2.5",
         "laravelcollective/html": "^6.2",
         "laravelcollective/remote": "^6.2",
-        "league/csv": "^9.7",
+        "league/flysystem-aws-s3-v3": "~1.0",
         "nesbot/carbon": "^2.46",
-        "phpmailer/phpmailer": "^6.4",
-        "phpoffice/phpspreadsheet": "^1.17",
+        "pbmedia/laravel-ffmpeg": "^7.5",
+        "socialiteproviders/facebook": "^4.1",
+        "socialiteproviders/google": "^4.1",
+        "socialiteproviders/vkontakte": "^4.2",
+        "tymon/jwt-auth": "^1.0",
+        "watson/rememberable": "^5.0",
         "yajra/laravel-datatables-buttons": "^4.13",
         "yajra/laravel-datatables-html": "^4.36",
         "yajra/laravel-datatables-oracle": "^9.18"
     },
     "require-dev": {
-        "facade/ignition": "^2.3.6",
-        "fzaninotto/faker": "^1.9.1",
-        "mockery/mockery": "^1.3.1",
+        "facade/ignition": "^2.5",
+        "fakerphp/faker": "^1.9.1",
+        "laravel/sail": "^1.0.1",
+        "mockery/mockery": "^1.4.2",
         "nunomaduro/collision": "^5.0",
-        "phpunit/phpunit": "^9.3"
-    },
-    "config": {
-        "optimize-autoloader": true,
-        "preferred-install": "dist",
-        "sort-packages": true
-    },
-    "extra": {
-        "laravel": {
-            "dont-discover": []
-        }
+        "phpunit/phpunit": "^9.3.3"
     },
     "autoload": {
         "psr-4": {
@@ -55,8 +48,6 @@
             "Tests\\": "tests/"
         }
     },
-    "minimum-stability": "dev",
-    "prefer-stable": true,
     "scripts": {
         "post-autoload-dump": [
             "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
@@ -68,5 +59,17 @@
         "post-create-project-cmd": [
             "@php artisan key:generate --ansi"
         ]
-    }
+    },
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true
+    },
+    "minimum-stability": "dev",
+    "prefer-stable": true
 }
