@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers\Admin;
 
-use App\Models\{Admin, Users};
+use App\Models\{Admin, User, Settings};
 use Illuminate\Support\Facades\Auth;
 use DataTables;
 use URL;
@@ -15,16 +15,14 @@ class DataTableController extends Controller
      */
     public function getUsers()
     {
-        $row = Users::query();
+        $row = User::query();
 
         return Datatables::of($row)
-            ->addColumn('action', function ($row) {
-                $editBtn = '';
-                $deleteBtn = '<a title="удалить" class="btn btn-xs btn-danger deleteRow" id="' . $row->id . '"><span class="fa fa-remove"></span></a>';
-
-                return '<div class="nobr"> ' . $editBtn . $deleteBtn . '</div>';
+            ->addColumn('checkbox', function ($row) {
+                return '<input type="checkbox" class="check" value="' . $row->id . '" name="status[]">';
             })
-            ->rawColumns(['action'])->make(true);
+
+            ->rawColumns(['checkbox'])->make(true);
     }
 
     /**
@@ -42,6 +40,23 @@ class DataTableController extends Controller
                     $deleteBtn = '<a title="удалить" class="btn btn-xs btn-danger deleteRow" id="' . $row->id . '"><span class="fa fa-remove"></span></a>';
                 else
                     $deleteBtn = '';
+
+                return '<div class="nobr"> ' . $editBtn . $deleteBtn . '</div>';
+            })
+            ->rawColumns(['action'])->make(true);
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getSettings()
+    {
+        $row = Settings::query()->dontRemember();
+
+        return Datatables::of($row)
+            ->addColumn('action', function ($row) {
+                $editBtn = '<a title="редактировать" class="btn btn-xs btn-primary"  href="' . URL::route('cp.settings.edit', ['id' => $row->id]) . '"><span  class="fa fa-edit"></span></a> &nbsp;';
+                $deleteBtn = '<a title="удалить" class="btn btn-xs btn-danger deleteRow" id="' . $row->id . '"><span class="fa fa-remove"></span></a>';
 
                 return '<div class="nobr"> ' . $editBtn . $deleteBtn . '</div>';
             })
