@@ -28,10 +28,10 @@ class Complaint extends Model
     ];
 
     /**
-     * @return mixed
+     * @return hasOne
      */
-    public function user()
+    public function user(): hasOne
     {
-        return $this->hasOne(Users::class,'id','user_id');
+        return $this->hasOne(User::class,'id','user_id');
     }
 }
