@@ -65,6 +65,7 @@ return [
 
         'pgsql' => [
             'driver' => 'pgsql',
+            'url' => env('DATABASE_URL'),
             'host' => env('DB_HOST', '127.0.0.1'),
             'port' => env('DB_PORT', '5432'),
             'database' => env('DB_DATABASE', 'forge'),
@@ -72,11 +73,9 @@ return [
             'password' => env('DB_PASSWORD', ''),
             'charset' => 'utf8',
             'prefix' => '',
-            'schema' => 'improover',
+            'prefix_indexes' => true,
+            'schema' => 'public',
             'sslmode' => 'prefer',
-            'options'   => array(
-                PDO::ATTR_PERSISTENT => true,
-            ),
         ],
 
         'sqlsrv' => [
