@@ -11,7 +11,7 @@ class CreateInterestsTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('interests', function (Blueprint $table) {
             $table->increments('id');
@@ -25,7 +25,7 @@ class CreateInterestsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('interests');
     }
