@@ -11,7 +11,7 @@ class CreateAdminTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('admin', function (Blueprint $table) {
             $table->increments('id');
@@ -28,7 +28,7 @@ class CreateAdminTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('admin');
     }
