@@ -11,7 +11,7 @@ class CreateSubscriptionsTable extends Migration
      *
      * @return void
      */
-    public function up()
+    public function up(): void
     {
         Schema::create('subscriptions', function (Blueprint $table) {
             $table->integer('user_id')->index('subscriptions_user_id');
@@ -25,7 +25,7 @@ class CreateSubscriptionsTable extends Migration
      *
      * @return void
      */
-    public function down()
+    public function down(): void
     {
         Schema::dropIfExists('subscriptions');
     }
