@@ -19,22 +19,54 @@
                 <div>
 
                     <div class="table-responsive">
+
+                        {!! Form::open(['url' => URL::route('cp.users.status'), 'method' => 'put']) !!}
+
                         <table id="itemList" class="table table-striped table-bordered table-hover" width="100%">
                             <thead>
                             <tr>
+                                <th width="10px">
+                                <span>
+                                    <input type="checkbox"  title="Отметить все/Снять отметку у всех" id="checkAll">
+                                </span>
+                                </th>
                                 <th>Фамилия</th>
                                 <th>Имя</th>
-                                <th>Отчество</th>
+                                <th>Никнейм</th>
                                 <th>Email</th>
-                                <th>Дата рождения</th>
                                 <th>Роль</th>
+                                <th>Дата рождения</th>
                                 <th>Дата регистрации</th>
-                                <th width="20px">Действия</th>
+                                <th>Удален</th>
+                                <th>Заблокирован</th>
                             </tr>
                             </thead>
                             <tbody>
                             </tbody>
                         </table>
+
+                        <div class="row">
+                            <div class="col-sm-12 padding-bottom-10">
+                                <div class="form-inline">
+                                    <div class="control-group">
+
+                                        {!! Form::select('action',[
+                                        '0' => 'заблокировать',
+                                        '1' => 'разблокировать',
+                                        '2' => 'сделать автором'
+                                        ],null,['class' => 'span3 form-control', 'id' => 'select_action','placeholder' => '--выберите действия--']) !!}
+
+                                        <span class="help-inline">
+                                           {!! Form::submit('применить', ['class' => 'btn btn-success', 'disabled' => "", 'id' => 'apply']) !!}
+                                        </span>
+
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+
+                        {!! Form::close() !!}
+
                     </div>
 
                 </div>
@@ -55,6 +87,38 @@
     <script>
 
         $(document).ready(function () {
+
+            $("#apply").click(function (event) {
+                var idSelect = $('#select_action').val();
+
+                if (idSelect == '') {
+                    event.preventDefault();
+                    swal({
+                        title: "Error",
+                        text: "выберите действия",
+                        type: "error",
+                        showCancelButton: false,
+                        cancelButtonText: "отменить",
+                        confirmButtonColor: "#DD6B55",
+                        closeOnConfirm: false
+                    });
+                } else {
+                    if (isConfirm) form.submit();
+                }
+            })
+
+            $("#checkAll").click(function () {
+                $('input:checkbox').not(this).prop('checked', this.checked);
+                countChecked();
+            });
+
+            $("#checkAll").on('change',function () {
+                countChecked();
+            });
+
+            $("#itemList").on('change', 'input.check', function () {
+                countChecked();
+            });
 
             pageSetUp();
 
@@ -113,55 +177,34 @@
                 'createdRow': function (row, data, dataIndex) {
                     $(row).attr('id', 'rowid_' + data['id']);
                 },
+                aaSorting: [[1, 'asc']],
                 processing: true,
                 serverSide: true,
                 ajax: {
                     url: '{{ URL::route('cp.datatable.users') }}'
                 },
                 columns: [
+                    {data: 'checkbox', name: 'checkbox', orderable: false, searchable: false},
                     {data: 'lastname', name: 'lastname'},
                     {data: 'firstname', name: 'firstname'},
-                    {data: 'secondname', name: 'secondname'},
+                    {data: 'nickname', name: 'nickname'},
                     {data: 'email', name: 'email'},
-                    {data: 'birthday', name: 'birthday'},
                     {data: 'role', name: 'role'},
+                    {data: 'birthday', name: 'birthday'},
                     {data: 'created_at', name: 'created_at'},
-                    {data: 'action', name: 'action', orderable: false, searchable: false},
+                    {data: 'deleted_at', name: 'deleted_at'},
+                    {data: 'banned_at', name: 'banned_at'},
                 ],
             });
-
-            $('#itemList').on('click', 'a.deleteRow', function () {
-
-                var rowid = $(this).attr('id');
-                swal({
-                        title: "Вы уверены?",
-                        text: "Вы не сможете восстановить эту информацию!",
-                        type: "warning",
-                        showCancelButton: true,
-                        confirmButtonColor: "#DD6B55",
-                        confirmButtonText: "Да",
-                        cancelButtonText: "Отмена",
-                        closeOnConfirm: false
-                    },
-                    function (isConfirm) {
-                        if (!isConfirm) return;
-                        $.ajax({
-                            url: SITE_URL + "/users/destroy/" + rowid,
-                            type: "DELETE",
-                            dataType: "html",
-                            headers: {'X-CSRF-TOKEN': $('meta[name="csrf-token"]').attr('content')},
-                            success: function () {
-                                $("#rowid_" + rowid).remove();
-                                swal("Сделано!", "Данные успешно удалены!", "success");
-                            },
-                            error: function (xhr, ajaxOptions, thrownError) {
-                                swal("Ошибка при удалении!", "Попробуйте еще раз", "error");
-                            }
-                        });
-                    });
-            });
-
         })
+
+        function countChecked()
+        {
+            if ($('.check').is(':checked'))
+                $('#apply').attr('disabled',false);
+            else
+                $('#apply').attr('disabled',true);
+        }
 
     </script>
 
