@@ -2,20 +2,9 @@
 
 use Illuminate\Support\Facades\Route;
 
-/*
-|--------------------------------------------------------------------------
-| Web Routes
-|--------------------------------------------------------------------------
-|
-| Here is where you can register web routes for your application. These
-| routes are loaded by the RouteServiceProvider within a group which
-| contains the "web" middleware group. Now create something great!
-|
-*/
+use App\Http\Controllers\Auth\LoginController;
 
-//отображение формы аутентификации
-Route::get('login', 'Auth\LoginController@showLoginForm')->name('login');
 
-//POST запрос аутентификации на сайте
-Route::post('login', 'Auth\LoginController@login')->name('singin');
-Route::get('logout', 'Auth\LoginController@logout')->name('logout');
+Route::get('login', [LoginController::class, 'showLoginForm'])->name('login');
+Route::post('login', [LoginController::class, 'login'])->name('singin');
+Route::get('logout', [LoginController::class, 'logout'])->name('logout');
